Move State to separate mod
